# portless-daemon/api/openapi.yaml
openapi: 3.1.0
info:
  title: Portless local control API
  version: 10.7.0
  description: >-
    Loopback-only API shared by the Portless CLI and embedded UI. Projects are
    reusable application topologies. Environments are independently configured
    and runnable instances. Public routes use readable project, environment,
    service, and artifact names; private ownership keys are never exposed.
    Authenticated native clients may send Portless-Client-Kind as `cli` or
    `mcp`; the daemon records that fixed actor category on durable operations.
    Browser sessions cannot claim a native client kind.
servers:
  - url: http://portless.localhost/api/v1
security:
  - bearer: []
  - browserSession: []
paths:
...
  /environments:
    get:
      operationId: listEnvironments
      parameters:
        - in: query
          name: project
          schema: { $ref: '#/components/schemas/ProjectName' }
        - { $ref: '#/components/parameters/Limit1000' }
      responses:
        '200':
          description: Environments across all projects or one filtered project
          content:
            application/json:
              schema:
                type: object
                required: [environments]
                properties:
                  environments:
                    type: array
                    items: { $ref: '#/components/schemas/Environment' }
    post:
      operationId: cloneEnvironment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [project, name]
              properties:
                project: { $ref: '#/components/schemas/ProjectName' }
                name: { $ref: '#/components/schemas/EnvironmentName' }
                from: { $ref: '#/components/schemas/EnvironmentName', default: local }
      responses:
        '201':
          description: New stopped environment with copied source and provider configuration
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Environment' }
...
  /environments/{projectName}/{environmentName}:
    parameters: &environmentParameters
      - { $ref: '#/components/parameters/ProjectName' }
      - { $ref: '#/components/parameters/EnvironmentName' }
    get:
      operationId: getEnvironment
      responses:
        '200':
          description: Effective sources, providers, topology, and runtime state
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Environment' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      operationId: forgetEnvironment
      responses:
        '204': { description: Stopped environment metadata forgotten }
...
  /environments/{projectName}/{environmentName}/services/{serviceName}/start:
    parameters: *serviceParameters
    post:
      operationId: startService
      parameters:
        - { $ref: '#/components/parameters/IdempotencyKey' }
      responses:
        '202':
          description: Idempotent service-start operation accepted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Operation' }
        '409': { $ref: '#/components/responses/Conflict' }
  /environments/{projectName}/{environmentName}/services/{serviceName}/stop:
    parameters: *serviceParameters
    post:
      operationId: stopService
      parameters:
        - { $ref: '#/components/parameters/IdempotencyKey' }
      responses:
        '202':
          description: Idempotent service-stop operation accepted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Operation' }
        '409': { $ref: '#/components/responses/Conflict' }
  /environments/{projectName}/{environmentName}/services/{serviceName}/restart:
    parameters: *serviceParameters
    post:
      operationId: restartService
      parameters:
        - { $ref: '#/components/parameters/IdempotencyKey' }
      responses:
        '202':
          description: Service-restart operation accepted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Operation' }
        '409': { $ref: '#/components/responses/Conflict' }
  /environments/{projectName}/{environmentName}/services/{serviceName}/manage:
    parameters: *serviceParameters
    post:
      operationId: manageService
      parameters:
        - { $ref: '#/components/parameters/IdempotencyKey' }
      responses:
        '202':
          description: Service restart in normal managed mode accepted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Operation' }
        '409': { $ref: '#/components/responses/Conflict' }
  /environments/{projectName}/{environmentName}/services/{serviceName}/debug:
    parameters: *serviceParameters
    post:
      operationId: debugService
      parameters:
        - { $ref: '#/components/parameters/IdempotencyKey' }
      responses:
        '202':
          description: Service restart with its discovered debugger enabled accepted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Operation' }
        '409': { $ref: '#/components/responses/Conflict' }
  /environments/{projectName}/{environmentName}/services/{serviceName}/configuration:
    parameters: *serviceParameters
    get:
      operationId: serviceConfiguration
      responses:
        '200':
          description: Effective configuration with sensitive values masked and generated connection values identified
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ServiceConfiguration' }
...
  /environments/{projectName}/{environmentName}/timeline:
    parameters: *environmentParameters
    get:
      operationId: timeline
      parameters:
        - { $ref: '#/components/parameters/Limit1000' }
      responses:
        '200':
          description: Newest-first durable environment timeline
          content:
            application/json:
              schema:
                type: object
                required: [timeline]
                properties:
                  timeline: { type: array, items: { $ref: '#/components/schemas/TimelineEvent' } }
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      description: Mode-0600 installation secret used by the local CLI.
    browserSession:
      type: apiKey
      in: cookie
      name: portless_session
      description: SameSite=Strict browser session; mutations also require X-Portless-CSRF.
  parameters:
...
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      description: >-
        Optional stable key for a durable lifecycle request. Reusing a key with
        the same request returns the original operation; reusing it for a
        different request returns IDEMPOTENCY_CONFLICT.
      schema: { type: string, maxLength: 200 }
    ProjectName:
      in: path
      name: projectName
      required: true
      schema: { $ref: '#/components/schemas/ProjectName' }
    EnvironmentName:
      in: path
      name: environmentName
      required: true
      schema: { $ref: '#/components/schemas/EnvironmentName' }
...
    Limit1000:
      in: query
      name: limit
      description: Positive result bound; the endpoint applies its documented default when omitted.
      schema: { type: integer, minimum: 1, maximum: 1000 }
...
  responses:
    NotFound:
      description: Named resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Conflict:
      description: Name, revision, provider, or current-state conflict
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
  schemas:
    ProjectName: &dnsName
      type: string
      pattern: '^[a-z][a-z0-9-]{0,62}$'
      examples: [billing]
...
    ArtifactName:
      type: string
      pattern: '^[a-z0-9][a-z0-9._-]{0,63}$'
      examples: [checkout-debug]
...
    SourceInput:
      type: object
      required: [name, path]
      properties:
        name: { $ref: '#/components/schemas/SourceName' }
        path: { type: string, description: Absolute path resolved by the CLI. }
...
    SourceBinding:
      allOf:
        - { $ref: '#/components/schemas/SourceInput' }
        - type: object
          required: [status, createdAt, scannedAt]
          properties:
            status: { type: string }
            warnings: { type: array, items: { type: string } }
            createdAt: { type: string, format: date-time }
            scannedAt: { type: string, format: date-time }
    RemoteTarget:
      type: object
      required: [url, classification, writePolicy]
      properties:
        url: { type: string, format: uri, pattern: '^https?://' }
        classification: { type: string, enum: [development, qa, staging, unknown] }
        writePolicy: { type: string, enum: [read-only, read-write] }
        healthPath: { type: string, pattern: '^/' }
    MockTarget:
      type: object
      required: [profile]
      properties:
        profile: { $ref: '#/components/schemas/ArtifactName' }
    ComponentBinding:
      type: object
      required: [provider]
      properties:
        service: { $ref: '#/components/schemas/ServiceName' }
        provider: { type: string, enum: [local, container, remote, mock] }
        source: { $ref: '#/components/schemas/SourceName' }
        remote: { $ref: '#/components/schemas/RemoteTarget' }
        mock: { $ref: '#/components/schemas/MockTarget' }
        modifiedAt: { type: string, format: date-time, readOnly: true }
      description: >-
        local requires source, container is valid only for discovered managed
        dependencies, remote requires remote configuration, and mock requires
        an environment-scoped profile attached to the same service.
...
    HealthCheck:
      type: object
      required: [kind, timeout, interval]
      properties:
        kind: { type: string, enum: [tcp, http, exec] }
        path: { type: string }
        timeout: { type: integer, description: Nanoseconds as encoded by Go time.Duration. }
        interval: { type: integer, description: Nanoseconds as encoded by Go time.Duration. }
    ServiceDefinition:
      type: object
      required: [name, kind, required, health]
      properties:
        name: { $ref: '#/components/schemas/ServiceName' }
        kind: { type: string, enum: [process, resource] }
        framework: { type: string }
        resource:
          type: object
          required: [type, version]
          properties:
            type: { type: string, example: postgres }
            version: { type: string, example: '17' }
        command: { type: array, items: { type: string } }
        workingDirectory: { type: string }
        serviceDirectory: { type: string }
        debug:
          type: object
          required: [adapter, launcher, command]
          properties:
            adapter: { type: string, enum: [node-inspector, jdwp] }
            launcher: { type: string, enum: [node-direct, nest-cli, spring-gradle, spring-maven] }
            command: { type: array, items: { type: string } }
        portEnvironment: { type: string }
        port: { type: integer, minimum: 1, maximum: 65535, description: Plugin-resolved client-facing port for a TCP service. }
        environment: { type: object, additionalProperties: { type: string } }
        required: { type: boolean }
        health: { $ref: '#/components/schemas/HealthCheck' }
    Service:
      allOf:
        - { $ref: '#/components/schemas/ServiceDefinition' }
        - type: object
          required: [status, launchMode, generation, endpoints, restartCount, recentRequests]
          properties:
            status: { type: string, enum: [planned, starting, recovering, ready, unhealthy, exited, failed, stopping, stopped, unknown] }
            launchMode: { type: string, enum: [managed, debug] }
            debugger:
              type: object
              required: [adapter, host, port, state]
              properties:
                adapter: { type: string, enum: [node-inspector, jdwp] }
                host: { type: string, const: '127.0.0.1' }
                port: { type: integer, minimum: 1, maximum: 65535 }
                state: { type: string, enum: [starting, listening, stopped] }
            reason: { type: string }
            generation: { type: integer, format: int64 }
            pid: { type: integer }
            upstreamPort: { type: integer }
            endpoints: { type: array, items: { $ref: '#/components/schemas/Endpoint' } }
            startedAt: { type: string, format: date-time }
            restartCount: { type: integer, format: int64 }
            recentRequests: { type: integer, format: int64 }
            p95Millis: { type: integer, format: int64 }
    Connection:
      type: object
      required: [source, target, protocol, required]
      properties:
        source: { type: string, description: Service name or external. }
        target: { $ref: '#/components/schemas/ServiceName' }
        protocol: { type: string, enum: [http, tcp] }
        binding: { type: string, description: Resource plugin responsible for connection formatting. }
        environment: { type: string }
        required: { type: boolean }
    Endpoint:
      type: object
      required: [kind, protocol, host, port, url]
      properties:
        kind: { type: string, enum: [public, connection] }
        protocol: { type: string, enum: [http, tcp] }
        host: { type: string }
        port: { type: integer, minimum: 1, maximum: 65535 }
        url: { type: string, format: uri, example: 'tcp://postgres.local.store.portless.test:5432' }
        address: { type: string, description: Internal listener address when needed for diagnostics. }
...
    ConfigurationValue:
      type: object
      required: [key, value, classification, source]
      properties:
        key: { type: string }
        value: { type: string }
        classification: { type: string, enum: [public, masked, generated] }
        source: { type: string }
    ServiceConfiguration:
      type: object
      required: [service, command, environment, health]
      properties:
        service: { $ref: '#/components/schemas/ServiceName' }
        command: { type: array, items: { type: string } }
        workingDirectory: { type: string }
        portEnvironment: { type: string }
        environment: { type: array, items: { $ref: '#/components/schemas/ConfigurationValue' } }
        health: { $ref: '#/components/schemas/HealthCheck' }
...
    EnvironmentStatus:
      type: string
      enum: [starting, recovering, healthy, degraded, failed, stopping, stopped, unknown]
    ConfigurationIssue:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        subject: { type: string }
        message: { type: string }
        remediation: { type: string }
    Environment:
      type: object
      required: [project, name, revision, status, createdAt, updatedAt, services, connections]
      properties:
        project: { $ref: '#/components/schemas/ProjectName' }
        name: { $ref: '#/components/schemas/EnvironmentName' }
        revision: { type: integer, format: int64 }
        status: { $ref: '#/components/schemas/EnvironmentStatus' }
        reason: { type: string }
        primaryService: { $ref: '#/components/schemas/ServiceName' }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        dashboardUrl: { type: string, format: uri }
        sources: { type: array, items: { $ref: '#/components/schemas/SourceBinding' } }
        bindings: { type: array, items: { $ref: '#/components/schemas/ComponentBinding' } }
        services: { type: array, items: { $ref: '#/components/schemas/Service' } }
        connections: { type: array, items: { $ref: '#/components/schemas/Connection' } }
        issues: { type: array, items: { $ref: '#/components/schemas/ConfigurationIssue' } }
...
    Operation:
      type: object
      required: [project, environment, number, type, state, actor, startedAt]
      properties:
        project: { $ref: '#/components/schemas/ProjectName' }
        environment: { $ref: '#/components/schemas/EnvironmentName' }
        number: { type: integer, format: int64 }
        type: { type: string }
        state: { type: string, enum: [running, succeeded, failed] }
        actor: { type: string }
        startedAt: { type: string, format: date-time }
        completedAt: { type: string, format: date-time }
        error: { type: string }
        events: { type: array, items: { $ref: '#/components/schemas/OperationEvent' } }
    OperationEvent:
      type: object
      required: [sequence, timestamp, type, message]
      properties:
        sequence: { type: integer, format: int64 }
        timestamp: { type: string, format: date-time }
        type: { type: string }
        subject: { type: string }
        message: { type: string }
        payload: { type: object, additionalProperties: true }
...
    TimelineEvent:
      type: object
      required: [project, environment, sequence, timestamp, actor, type, severity, summary]
      properties:
        project: { $ref: '#/components/schemas/ProjectName' }
        environment: { $ref: '#/components/schemas/EnvironmentName' }
        sequence: { type: integer, format: int64, minimum: 1 }
        timestamp: { type: string, format: date-time }
        actor: { type: string }
        type: { type: string }
        subject: { type: string }
        severity: { type: string, enum: [info, warning, error] }
        summary: { type: string }
        details: { type: object, additionalProperties: true }
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            subject: { type: object, additionalProperties: true }
            details: { type: object, additionalProperties: true }
            remediation: { type: array, items: { type: object, additionalProperties: true } }